+++ server/conference/__init__.py
@@ -1,4 +1,4 @@
-from server.conference.conference_manager import ConferenceManager
+from server.conference.conferences_pool import ConferencesPool
 from server.conference.message_coding.json_message_coder import JSONMessageCoder
 from server.config import CONFERENCE_MESSAGE_CONTRACT
 
@@ -8,4 +8,4 @@ match CONFERENCE_MESSAGE_CONTRACT:
     case _:
         main_message_coding = JSONMessageCoder()
 
-main_conference_manager = ConferenceManager(message_coding=main_message_coding)
+main_conferences_pool = ConferencesPool(message_coding=main_message_coding)
