+++ server/domains/conferences.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, WebSocket
 from fastapi.responses import JSONResponse, Response
 
-from server.conference import main_conference_manager
+from server.conference import main_conferences_pool
 from server.conference.exceptions import ConferenceNotFound
 from server.conference.member_connections import WebsocketMemberConnection
 from server.conference.types import ConferenceID
@@ -15,7 +15,7 @@ _id = None
 @router.post("/conference/")
 async def post():
     global _id
-    conference_controller = main_conference_manager.create_conference()
+    conference_controller = main_conferences_pool.create_conference()
     _id = conference_controller.conference_id
     return JSONResponse({"conference_id": str(conference_controller.conference_id)})
 
@@ -24,7 +24,7 @@ async def post():
 async def websocket_endpoint(conference_id: str, websocket: WebSocket):
     connection = WebsocketMemberConnection(websocket)
     try:
-        conference_controller = main_conference_manager.get_conference(
+        conference_controller = main_conferences_pool.get_conference(
             ConferenceID(conference_id)
         )
     except ConferenceNotFound:
